fix(knowledge): split reranker model constants into client-safe module
